requirements.txt
fastapi
starlette
uvicorn
starlette
redis
ray
pyarrow
getdaft
lancedb
mlflow
openai
vllm
outlines
numpy
pandas
torch
transformers
pydantic
boto3
google-api-python-client
azure-cli
azure-core 
jsonschema 
jsonschema_specifications